connect a component to redux and verify RHL still works
