+++ src/containers/app/App.css
@@ -1,0 +1,3 @@
+.app {
+  text-align: center;
+}
