+++ src/containers/home/home.css
@@ -1,7 +1,3 @@
-.app {
-  text-align: center;
-}
-
 .logo {
   animation: logo-spin infinite 20s linear;
   height: 80px;
