+++ src/index.dev.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 
 import { configureStore } from './redux/store'
-import App from './App'
+import { App } from './containers/app'
 import './index.css'
 
 const store = configureStore()
