+++ src/index.prod.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './index.css'
-import App from './App'
+import { App } from './containers/app'
 import registerServiceWorker from './registerServiceWorker'
+import './index.css'
 
 ReactDOM.render(<App />, document.getElementById('root'))
 registerServiceWorker()
